terminando

+++ src/main/java/org/models/Aviao.java
@@ -2,9 +2,9 @@ package org.models;
 
 public class Aviao extends Veiculo{
     private double velocidadecruzeiro;
-    private int tempodevoo;
+    private double tempodevoo;
 
-    public Aviao(String identificacao,double velocidadecruzeiro,int tempodevoo, double valordecompra){
+    public Aviao(String identificacao,double velocidadecruzeiro,double tempodevoo, double valordecompra){
         super(identificacao,valordecompra);
         this.velocidadecruzeiro=velocidadecruzeiro;
         this.tempodevoo=tempodevoo;
